API
Added Stamp and Mention to the API. Handling of Stamp is notable: for the function getStamp, I'm pulling in data for User and Entity via separate calls to the database. Arguably this is correct for scaling, even though it goes against all of those SQL best practices :)

Still need to add some custom functions (e.g. marking a stamp as read) and finish out the rest of the objects.