API: Adding Collections, Favorites, Followers, Friends, and Friendships
I carved out Followers and Friends (lists of ids), even though arguably they belong in Friendships. We can discuss that -- for now I wanted to copy the structure of our eventually-public API (which carves them out a la Twitter).

This is almost everything. I still need to hit Account and Blocks, as well as the inbox in Collections, matching in Entities, and flagging in Stamps, Comments, and Entities. (And Linked Accounts, but that's flat-out being ignored for now.) I think I also need to take another look at Users.

One question I have is how the structure should exist for something like a mention. Currently we treat that as a separate object, when arguably it's a property of a stamp. I've identified a handful of things that are STAMP::USER (i.e. owned by the stamp but reference users, such as mentions) and USER::STAMP (i.e. owned by the user but referencing stamps, such as favorites). Not sure how we want to handle these ultimately, but they're in there as separate objects for now.