API: Blocks
Added Blocks, and added the proper check for blocks and for privacy before a friend can be added. Where we should run this check is debatable, but I figured I would add it here for now (easy to take out later).

This is most of the API. I still need to carve out Accounts from Users, but I'm not yet certain how that should happen, so I'm waiting for the moment. I also need to build an interface _to_ the API -- right now all you can do is run a test that goes through everything, but you can't yet call functions. I either need to figure out the optparser myself or work with Travis to get that implemented properly.