Splash Page v2
First cut of Bons' version of the splash page. Mostly working -- input is stylized and properly checks for completion. Big issue is that asynchronous submit doesn't work on Firefox (works properly in Chrome / Safari). Need to figure out what's going on...

Also includes work-in-progress Mongo files for the API. More on that later.