Splash Page
Updated the splash page. This should now be launch-ready, but we should double-check it to make sure there are no silly errors.

Note: The Firefox bug I ran into earlier involves the same origin policy (http://en.wikipedia.org/wiki/Same_origin_policy) which prevents the site from using AJAX to load data taken from Google's domain. This poses a problem given that we're posting to Google Docs in order to save the data. There appear to be two suggested methods for getting around this:

1) Post to a local page asynchronously that then forwards the request to the third-party site. This requires setting up a back-end script and adding any proper validation of the content. Shouldn't be too hard to do, but I wasn't sure of how Robby's dreamhost server is set up and didn't want to mess around too much getting this working.

2) Use the JSONP (http://en.wikipedia.org/wiki/JSONP) work-around by dynamically sending Google the callback function (http://code.google.com/webtoolkit/doc/latest/FAQ_Server.html and http://code.google.com/apis/gdata/docs/json.html). I tried this briefly and didn't spend much time with it.

The "solution" I settled upon was to ignore the problem -- i.e. to ignore any validation that the script was received and just display the thank you message upon submission. This means we could potentially miss requests that are sent and the user wouldn't know it, but that's a risk I'm willing to take in the short term (i.e. next few days). We should change this before too long.