Mongo API - Entity, User, and Stamp
First cut of the API for Entity, User, and Stamp. Only basic CRUD actions at the moment, with simple hooks for a few other things (returning entities based on string matching and returning users by id, username, or username string matching). It includes basic validation for Users and Stamps but is not yet fleshed out for Entities.

Note that I'm converting the MongoDB _id (which is created as an object in PyMongo -- see http://api.mongodb.org/python/current/api/bson/objectid.html) to a basestring id that's presented to each object. This is mostly to ensure that the API will function normally if we switch to a database that doesn't treat ids as objects. (We'll also likely treat ids as a string on the client side, so it's easiest to do the conversion here.)

Also note that we'll probably break up User into two different objects (User and Account), making User read-only and forcing all write operations to happen via Account. This is to help ensure that no one can accidentally access password information via the User class, since that will be accessed less stringently than the Account class. That said, both will likely exist as the same document within MongoDB.

MongoDB.py is where most of the magic happens, and '_mapDataToSchema' will probably need more validation at some point. It should work just fine for now.

If you want to run api.py it will run through a handful of tests currently set up. Note that this is accessing the AWS cluster we have set up, so some configuration may need to be changed ultimately.