API - Favorites
Added Favorites (aka To Do List). This required both the object / collection MongoFavorites and the relationship MongoUserFavorites. Note that we're storing the stamp referral as well as the creator of the stamp within a Favorite so that it's easy to reference in the Entity Detail view when the user views it from their To Do List; this is also in case the original stamp is deleted.

This leaves Conversation, Blocks, and Account as the major remaining collections to build. We also need to do Directed, Credit, and Mentions, but the requirements for those collections are still a bit nebulous, so I'm placing them as a lower priority.