API - Unit Tests, Documentation, and Reformatting
Started adding documentation to the API, as well as modifying the struct
ure of the HTTP interface to more closely mirror the finished product. M
ostly done with Accounts, Users, and Entities at this point (although th
ere are a few functions which do not work perfectly yet). First draft of
 the documentation for Stamps, Friendship, and Collections, but still ne
eds to be fully implemented in the API itself.