API - Added "since" and "before" to Inbox
- I added "before" and "since" as fields to collections/inbox.json. If y
ou pass them a unix timestamp, it will use that to slice the data approp
riately. By default they're empty, i.e. "since" is the beginning of the
inbox and "before" is now. An example query is http://192.168.0.10:5000/
api/v1/collections/inbox.json?authenticated_user_id=4e2792f732a7ba6a5600
04b1&limit=10&since=1311216375&before=1311216377. Note that I'm actually
 subtracting two seconds from "since" and adding two seconds to "before"
 to make the gap a bit larger, just because Mongo's timestamp has some v
ariability in there and I want to make sure we're not accidentally exclu
ding data.

- I modified "limit" for collection/inbox.json as well, so i
t should be working as expected. It defaults to 20, and if you pass it a
 number that's less than 20 it will use that instead.

- I created a new
 generic function in serve.py called handleRequest. It basically works e
xactly the same as handlePOSTRequest, but you pass it the MultiDict of a
rguments (either request.form or request.args) as opposed to request, so
 that it will work for either POST or GET. I think I might change this s
o that it just takes the request and then pulls out form or args based o
n request.method, but am still testing it for now. I'm using it for coll
ections/inbox.json since I needed to be able to pass optional arguments.