API - Cleanup
Updated serve.py to only use handleRequest() and phased out
handleGETRequest() and handlePOSTRequest(). All functions have been
tested and appear to be working as expected. I also tested the "return"
functionality in StampedAPI.py that I added last night and verified that
everything is working as expected.   Still need to add documentation for
limits and slices, and we still need better error handling and
validation within StampedAPI.py, but it's a step in the right direction.