MongoDB - Activity
Added first cut of activity. This is populated by adding a stamp or a
comment. I've not yet added the API hooks to access the activity feed,
but that should be a pretty quick add tomorrow. I also still have to
look at activity population from favorites, directed (on hold), and
credit milestones (also TBD).