OAuth 2.0 v1
Rolling out v1 of OAuth 2.0 implementation in Stamped. The juicy
details:  - Upon registration of a user, the client will receive an
access token and a refresh token (in addition to the validated account
details). The access token is valid for a limited amount of time
(denoted in the response, currently set to just over one hour) and is
used as a shared secret to allow the client to make requests to the API.
- A refresh token and access token can also be obtained by logging in
the user via "account/login.json". Currently this only accepts the
screen name, although eventually it should support email as well. Note
that this may ultimately have an option as to whether or not to generate
a refresh token, or it may be set based upon the client id -- that's all
TBD.  - Speaking of client ids, that's currently hardcoded to the same
auto we had before ("stampedtest" and "august1ftw"). We can change those
to something more secure later on.  - The client can refresh its access
id by passing the client id, client secret, and refresh token to
"oauth2/token.json". See TestAPI.py for an example (in the Collections
section for now).   - If anything breaks, you should be able to get away
with rolling back serve.py -- that's really the only file that changed.
(Everything else should flow through StampedAuth.py.)   We'll discuss
tomorrow in more detail….