Updates to Schema Work
Major updates -- finishing up first pass of StampedAPI.py. Favorites is
still incomplete, and Activity needs to be fully implemented (in
addStamp, etc.). Collections is cleaned up a bit and allows for the new
"quality" parameter to be passed (allowing for more results returned).
Still need to find the proper way to handle conversion of data,
especially for exports. Activity is a great example of the problems that
can arise. Will figure that out tomorrow!