Add initial support for derivatives. See notes.
First cut of deriving. If you pass a SchemaElement two
parameters("derivedFrom" and "derivedFn"), it will use the derivedFn to
calculate the value based on the value of the "derivedFrom" field. Note
that "derivedFrom" must currently be a sibling attribute, e.g. it cannot
be in a parent tree or in a sibling tree. I think I can get it working
properly if it's in a child element, but not sure if we want to support
looking at parent trees.   Also made a correction to the HTTPStamp
schema for Bons, posting so that we can update the test site.